chore: pin versions for supported plugins
Many plugins publish Git tags to allow users to specify which version to
install. For these supported plugins, I've supplied a SemVer range that
will tell `lazy.nvim` to only install patch updates. This way, breaking
changes won't be accidentally applied.

I have NOT pinned the version of plugins that are dependencies of other
plugins. This is to avoid version mismatches when a plugin requires a
certain version of a dependency. Additionally, some plugins don't
publish Git tags, so it is unfortunately not possible to pin their
versions.

Instead of manually managing the versions of plugins, it might be worth
it to take a look into `rocks.nvim`. It automatically manages plugin
dependencies, among other useful features.