chore(editor): link missing highlights
There seems to be a discrepancy between TypeScript's `tree-sitter`
syntax capture groups and those supported by Kanagawa. This may be
because the latest version of `nvim-treesitter` has renamed some capture
groups, which `tree-sitter-typescript` correctly adheres to, but
Kanagawa (and many other color schemes) do not.

As a result, some keywords (such as `import` and `if`) don't have any
highlighting, which affects readability. To resolve this, I've added
some manual highlight links for TypeScript and TypeScript React as
suggested in the Reddit thread (linked below). It may also be a good
idea to port all the changes suggested in the top comment in the thread.

https://www.reddit.com/r/neovim/comments/19aratu/psa_nvimtreesitter_breaking_changes_on_highlight/
https://github.com/nvim-treesitter/nvim-treesitter/pull/5895
https://github.com/nvim-treesitter/nvim-treesitter/commit/1ae9b0e4558fe7868f8cda2db65239cfb14836d0