fix(plugin): `nvim-treesitter` fails to work with Neovim v0.11.1
Occassionally, treesitter functionality stops working when I upgrade
Neovim and treesitter with Homebrew. This might be because I've
hard-coded the version of this plugin, which may not be compatible with
the newer versions of Neovim.

I'll need to make sure to manually update this plugin whenever I upgrade
Neovim and treesitter again.